+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue preventing building projects containing a "Play Music" event but no music
 - Fix issue where dialogue script events could cause horizontal scroll bars to appear in script editor when column was not wide enough to display all tabs
 - Fix issue where errors causing the build process to end early where not being display correctly in the Build Log
+- Fix issue preventing build when a scene includes a common tileset that has been deleted
 
 ## [4.1.2] - 2024-09-09